+++ src/common/config/index.ts
@@ -28,6 +28,8 @@ const ACCESS_TOKEN = (window as any)._env_.ACCESS_TOKEN;
 const SERVED_ENTITY_TYPES = (window as any)._env_.SERVED_ENTITY_TYPES;
 const PROJECT_VERSION = (window as any)._env_.PROJECT_VERSION;
 const WHATSNEW_URL = (window as any)._env_.WHATSNEW_URL;
+const SITES_CONFIG = JSON.parse((window as any)._env_.SITES_CONFIG);
+const BFF_PATH = (window as any)._env_.BFF_PATH;
 
 const enrichBaseMaps = (baseMaps: IBaseMaps): IBaseMaps => {
   return {
@@ -177,6 +179,7 @@ const APP_CONFIG = {
   JOB_MANAGER_END_OF_TIME: 21, // Days
   MINIMUM_SUPPORTED_BROWSER_VERSION: 84,
   PROJECT_VERSION: PROJECT_VERSION,
+  POLYGON_PARTS_FEATURE_TYPE_PREFIX: 'polygonParts:',
   CONTEXT_MENUS: {
     MAP: {
       MAX_ACTIVE_LAYERS_TO_PRESENT: 5,
@@ -187,7 +190,9 @@ const APP_CONFIG = {
       }
     }
   },
-  WHATSNEW_URL: WHATSNEW_URL
+  WHATSNEW_URL: WHATSNEW_URL,
+  SITES_CONFIG: SITES_CONFIG,
+  BFF_PATH: BFF_PATH
 };
 
 export default APP_CONFIG;