+++ src/discrete-layer/components/layer-details/layer-datails-form.tsx
@@ -1,5 +1,5 @@
 import React, { useMemo, useState, useEffect, useCallback } from 'react';
-import { FormattedMessage } from 'react-intl';
+import { FormattedMessage, useIntl } from 'react-intl';
 import {
   withFormik,
   FormikProps,
@@ -13,8 +13,11 @@ import { OptionalObjectSchema, TypeOfShape } from 'yup/lib/object';
 import { AnyObject } from 'yup/lib/types';
 import { DraftResult } from 'vest/vestResult';
 import { get, isEmpty } from 'lodash';
-import { Button } from '@map-colonies/react-core';
+import { Button, Checkbox, IconButton, Tooltip, Typography } from '@map-colonies/react-core';
 import { Box } from '@map-colonies/react-components';
+import { SYNC_QUERY_NAME } from '../../../syncHttpClientGql';
+import { emphasizeByHTML } from '../../../common/helpers/formatters';
+import { sessionStore } from '../../../common/helpers/storage';
 import { Mode } from '../../../common/models/mode.enum';
 import { ValidationsError } from '../../../common/components/error/validations.error-presentor';
 import { GraphQLError } from '../../../common/components/error/graphql.error-presentor';
@@ -34,6 +37,8 @@ import {
   extractDescriptorRelatedFieldNames,
   getFlatEntityDescriptors,
   transformEntityToFormFields,
+  getValidationMessage,
+  ValidationMessage,
 } from './utils';
 
 import './layer-details-form.css';
@@ -44,6 +49,12 @@ const NONE = 0;
 export interface FormValues {
   directory: string;
   fileNames: string;
+}
+
+export interface StatusError {
+  errors: {
+    [fieldName: string]: string[];
+  }
 }
 
 interface LayerDetailsFormCustomProps {
@@ -105,12 +116,14 @@ const InnerForm = (
   } = props;
 
   const status = props.status as StatusError | Record<string, unknown>;
-
+  const intl = useIntl();
   const [graphQLError, setGraphQLError] = useState<unknown>(mutationQueryError);
   const [isSelectedFiles, setIsSelectedFiles] = useState<boolean>(false);
   const [firstPhaseErrors, setFirstPhaseErrors] = useState<Record<string, string[]>>({});
   const [showCurtain, setShowCurtain] = useState<boolean>(true);
   const [gpkgValidationError, setGpkgValidationError] = useState<string|undefined>(undefined);
+  const [syncAnywayChecked, setSyncAnywayChecked] = useState<boolean>(false);
+  const [validationWarn, setValidationWarn] = useState<ValidationMessage>();
 
   const getStatusErrors = useCallback((): StatusError | Record<string, unknown> => {
     return {
@@ -155,6 +168,54 @@ const InnerForm = (
       ...getStatusErrors() as { [fieldName: string]: string[]; },
     })
   }, [errors, getYupErrors, getStatusErrors]);
+
+  useEffect(() => {
+    // Add method wathers for storage changes
+    sessionStore.watchMethods(
+      ['setItem', 'removeItem'],
+      (method, key, ...args) => {},
+      (method, key, ...args) => {
+        switch (true) {
+          case key.includes(SYNC_QUERY_NAME.GET_PRODUCT): {
+            switch (method) {
+              case 'setItem': {
+                const invalidVersion = sessionStore.getObject(SYNC_QUERY_NAME.GET_PRODUCT);
+                if (invalidVersion) {
+                  setValidationWarn(getValidationMessage(invalidVersion, intl));
+                }
+                break;
+              }
+              case 'removeItem': {
+                setValidationWarn(undefined);
+                break;
+              }
+            }
+            break;
+          }
+          case key.includes(SYNC_QUERY_NAME.VALIDATE_SOURCE): {
+            switch (method) {
+              case 'setItem': {
+                const sourceValidation = sessionStore.getObject(SYNC_QUERY_NAME.VALIDATE_SOURCE);
+                if (sourceValidation && !sourceValidation.isValid && !sessionStore.getObject(SYNC_QUERY_NAME.GET_PRODUCT)) {
+                  setValidationWarn(getValidationMessage(sourceValidation, intl));
+                }
+                break;
+              }
+              case 'removeItem': {
+                setValidationWarn(undefined);
+                break;
+              }
+            }
+            break;
+          }
+        }
+      }
+    );
+    // Clean up the method wathers when the component unmounts
+    return () => {
+      sessionStore.unWatchMethods();
+    };
+  }, []);
 
   const entityFormikHandlers: EntityFormikHandlers = useMemo(
     () => ({
@@ -239,10 +300,12 @@ const InnerForm = (
           <IngestionFields
             formik={entityFormikHandlers}
             reloadFormMetadata={reloadFormMetadata}
-            validateSources={true}
+            validateSources={false}
             recordType={recordType}
             fields={ingestionFields}
             values={values}
+            isError={showCurtain}
+            onErrorCallback={setShowCurtain}
           />
         }
         <Box
@@ -263,18 +326,55 @@ const InnerForm = (
           <Box className="messages">
             {
               Object.keys(firstPhaseErrors).length > NONE &&
+              JSON.stringify(firstPhaseErrors) !== '{}' &&
               <ValidationsError errors={firstPhaseErrors} />
             }
             {
-              Object.keys(errors).length === NONE &&
+              (Object.keys(errors).length === NONE || JSON.stringify(errors) === '{}') &&
               vestValidationResults.errorCount > NONE &&
               <ValidationsError errors={vestValidationResults.getErrors()} />
             }
             {
-              graphQLError !== undefined && (
-                // eslint-disable-next-line
-                <GraphQLError error={graphQLError} />
-              )
+              graphQLError !== undefined &&
+              graphQLError !== null &&
+              graphQLError &&
+              JSON.stringify(graphQLError) !== '{}' &&
+              Object.keys(graphQLError).length > NONE &&
+              <GraphQLError error={graphQLError} />
+            }
+            {
+              validationWarn?.message &&
+              isEmpty(firstPhaseErrors) &&
+              (!isEmpty(errors) || !vestValidationResults.errorCount) &&
+              isEmpty(graphQLError) &&
+              <Box className="ingestionWarning">
+                <Typography tag="span"><IconButton className="mc-icon-Status-Warnings warningIcon warning" /></Typography>
+                <Box>
+                  <Typography tag="div" className="ingestionWarningMessage">
+                    <Typography tag="span" className="warningMessage warning"
+                      dangerouslySetInnerHTML={{__html:
+                        intl.formatMessage(
+                          { id: 'ingestion.warning.invalid-secondary-file' },
+                          { title: emphasizeByHTML(`${intl.formatMessage({ id: 'ingestion.warning.title' })}`) }
+                        )
+                      }}
+                    />
+                    <Typography tag="span" className="warning">{' - '}</Typography>
+                    <Tooltip content={validationWarn?.message}>
+                      <Typography tag="span" className={validationWarn?.severity}>{validationWarn?.message}</Typography>
+                    </Tooltip>
+                  </Typography>
+                  <Checkbox
+                    className="warning"
+                    label={intl.formatMessage({id: 'ingestion.checkbox.label'})}
+                    checked={syncAnywayChecked}
+                    onClick={
+                      (evt: React.MouseEvent<HTMLInputElement>): void => {
+                        setSyncAnywayChecked(evt.currentTarget.checked);
+                      }}
+                  />
+                </Box>
+              </Box>
             }
           </Box>
           <Box className="buttons">
@@ -286,7 +386,8 @@ const InnerForm = (
                 !dirty ||
                 Object.keys(errors).length > NONE ||
                 (Object.keys(getStatusErrors()).length > NONE) ||
-                !isEmpty(graphQLError)
+                !isEmpty(graphQLError)  ||
+                (!isEmpty(validationWarn?.message) && !syncAnywayChecked)
               }
             >
               <FormattedMessage id="general.ok-btn.text" />