+++ src/discrete-layer/views/components/app-title/app-title.component.tsx
@@ -3,6 +3,7 @@ import { useIntl } from 'react-intl';
 import { Icon, Tooltip, Typography } from '@map-colonies/react-core';
 import { Box } from '@map-colonies/react-components';
 import CONFIG from '../../../../common/config';
+import { currentSite } from '../../../../common/helpers/siteUrl';
 import packageJson from '../../../../../package.json';
 
 
@@ -19,22 +20,21 @@ const AppTitle = (): JSX.Element => {
 
   return (
     <>
-    <Box className='appLogoContainer'>
-      <Box className='appTitleVersionContainer'>
-        <Typography tag="b">
+      <Box className='appLogoContainer'>
+        <Box className='appTitleVersionContainer'>
+          <Typography tag="b">
             {intl.formatMessage({ id: 'general.logo.text' })}
-        </Typography>
-        <Tooltip
-          content={`${intl.formatMessage({
-            id: 'general.version.tooltip',
-          })} ${projectVersion}`}
-        >
-          <Box className="appVersionContainer" dir='auto'>{versionText}</Box>
-      </Tooltip>
+          </Typography>
+          <Tooltip
+            content={`${intl.formatMessage({
+              id: 'general.version.tooltip',
+            })} ${projectVersion}`}
+          >
+            <Box className={`appVersionContainer-${currentSite()}`}> {versionText} </Box>
+          </Tooltip>
+        </Box>
+        <Icon className="appIcon mc-icon-AppIcon" />
       </Box>
-      <Icon className="appIcon mc-icon-AppIcon" />
-    </Box>
-     
     </>
   );
 };